v4.2.0 ======
Release Summary
---------------

New features (Port Security, PoE, MAC, static MAC and speed/duplex) and bugfixes

Major Changes
-------------

- Add new modules for PoE (aoscx_poe), MAC (aoscx_mac) and Static MAC (aoscx_static_mac)
- Add port security support (aoscx_l2_interface).
- Add speed and duplex support (aoscx_interface).
- Fix module aoscx_upload firmware using HTTP.
- Fix module aoscx_upload_firmware for local path (https://github.com/aruba/aoscx-ansible-collection/issues/28).
- Include fixes for issues found internally.

Minor Changes
-------------

- Fix Interface MTU support (https://github.com/aruba/aoscx-ansible-collection/issues/38).
- Fix idempotency in ACL module

+++ plugins/modules/aoscx_l2_interface.py
@@ -23,7 +23,8 @@ short_description: >
   Create or Update or Delete Layer2 Interface configuration on AOS-CX.
 description: >
   This modules provides configuration management of Layer2 Interfaces on AOS-CX
-  devices.
+  devices, including Port Security features. For platform 8360, Port Security
+  is supported from REST v10.09 upwards.
 author: Aruba Networks (@ArubaNetworks)
 options:
   interface:
@@ -217,14 +218,27 @@ EXAMPLES = """
 
 - name: >
     Configure Interface 1/1/13 - remove allowed MAC address AA:BB:CC:DD:EE:FF
-    Previous allowed MAC addresses:
-    - 12:34:56:78:90:01
-    - AA:BB:CC:DD:EE:FF
   aoscx_l2_interface:
     name: 1/1/13
     port_security_enable: true
     port_security_macs:
-      - 12:34:56:78:90:01
+      - AA:BB:CC:DD:EE:FF
+    state: delete
+
+- name: Configure Interface 1/1/13 - delete configuration of client limit.
+  aoscx_l2_interface:
+    name: 1/1/13
+    port_security_enable: true
+    port_security_client_limit: 2
+    state: delete
+
+- name: >
+    Configure Interface 1/1/13 - delete configuration of recovery time.
+  aoscx_l2_interface:
+    name: 1/1/13
+    port_security_enable: true
+    port_security_recovery_time: 60
+    state: delete
 
 - name: Configure Interface 1/1/13 - disable port security.
   aoscx_l2_interface:
@@ -287,7 +301,6 @@ EXAMPLES = """
     interface: 1/1/4
     vlan_mode: access
     vlan_access: '200'
-
 - name: >
     Configure Interface 1/1/5 - vlan trunk allowed all, vlan trunk native 200
     tag.
@@ -483,7 +496,7 @@ def main():
 
         session = get_pyaoscx_session(ansible_module)
         device = Device(session)
-        if state == "delete":
+        if state == "delete" and port_security_enable is None:
             interface = device.interface(interface_name)
             interface.delete()
 
@@ -510,48 +523,89 @@ def main():
             native_vlan_tag=native_vlan_tag,
         )
 
-        if port_security_enable is not None:
-            port_security_sticky_macs = port_security_sticky_macs or []
-            if not port_security_enable:
-                modified_op |= interface.port_security_disable()
+        if port_security_enable is not None and not port_security_enable:
+            modified_op |= interface.port_security_disable()
+        if port_security_enable or (
+            hasattr(interface, "port_security")
+            and interface.port_security["enable"]
+        ):
+            port_sec_kw = {}
+            if state == "delete":
+                if port_security_client_limit:
+                    port_sec_kw["client_limit"] = 1
+                if port_security_sticky_learning is not None:
+                    port_sec_kw["sticky_mac_learning"] = False
+                if port_security_macs:
+                    for mac in port_security_macs:
+                        mac = mac.upper()
+                        sw_static_macs = (
+                            interface.port_security_static_client_mac_addr
+                        )
+                        if mac in sw_static_macs:
+                            sw_static_macs.remove(mac)
+                            modified_op |= interface.apply()
+                        else:
+                            ansible_module.fail_json(
+                                msg="MAC address {0} is not configured".format(
+                                    mac
+                                )
+                            )
+                if port_security_sticky_macs:
+                    for sticky_mac in port_security_sticky_macs:
+                        mac = sticky_mac["mac"]
+                        sw_sticky_macs = (
+                            interface.port_security_static_sticky_client_mac_addr
+                        )
+                        sw_sticky_mac_vlans = sw_sticky_macs[mac]
+                        if mac in sw_sticky_macs:
+                            for vlan in sticky_mac["vlans"]:
+                                if vlan in sw_sticky_mac_vlans:
+                                    sw_sticky_mac_vlans.remove(vlan)
+                            if sw_sticky_mac_vlans == []:
+                                del sw_sticky_macs[mac]
+                        else:
+                            ansible_module.fail_json(
+                                msg="MAC address {0} is not configured".format(
+                                    mac
+                                )
+                            )
+                    modified_op |= interface.apply()
+                if port_security_violation_action:
+                    port_sec_kw["violation_action"] = "notify"
+                if port_security_recovery_time:
+                    port_sec_kw["violation_recovery_time"] = 10
             else:
-                port_sec_kw = {}
                 if port_security_client_limit:
                     port_sec_kw["client_limit"] = port_security_client_limit
-
                 if port_security_sticky_learning is not None:
                     port_sec_kw[
                         "sticky_mac_learning"
                     ] = port_security_sticky_learning
-
                 if port_security_macs:
                     port_sec_kw["allowed_mac_addr"] = port_security_macs
-
-                converted_sticky_macs = {
-                    el["mac"]: el["vlans"] for el in port_security_sticky_macs
-                }
-                if converted_sticky_macs:
+                if port_security_sticky_macs:
+                    converted_sticky_macs = {
+                        el["mac"]: el["vlans"]
+                        for el in port_security_sticky_macs
+                    }
                     port_sec_kw[
                         "allowed_sticky_mac_addr"
                     ] = converted_sticky_macs
-
                 if port_security_violation_action:
                     port_sec_kw[
                         "violation_action"
                     ] = port_security_violation_action
-
                 if port_security_recovery_time:
                     port_sec_kw[
                         "violation_recovery_time"
                     ] = port_security_recovery_time
+            _result = False
+            try:
+                _result = interface.port_security_enable(**port_sec_kw)
+            except Exception as exc:
+                ansible_module.fail_json(msg=str(exc))
 
-                _result = False
-                try:
-                    _result = interface.port_security_enable(**port_sec_kw)
-                except Exception as exc:
-                    ansible_module.fail_json(msg=str(exc))
-
-                modified_op |= _result
+            modified_op |= _result
         if modified_op:
             result["changed"] = True
 
@@ -665,7 +719,7 @@ def main():
                 aruba_ansible_module, interface_name, interface_vlan_dict
             )
 
-        if state == "delete":
+        if state == "delete" and port_security_enable is None:
             aruba_ansible_module = l2_interface.delete_l2_interface(
                 aruba_ansible_module, interface_name
             )
